user list done, theme updated, sweet aleert added

+++ components/ui/UserList.jsx
@@ -1,11 +1,33 @@
+import Button from "./Button";
+import Swal from 'sweetalert2';
+
 const UserList = ({ users, onDelete }) => {
+
     const handleDelete = async (id) => {
-      try {
-        await fetch(`/api/user/${id}`, { method: 'DELETE' });
-        onDelete(id);
-      } catch (error) {
-        console.error('Error deleting user:', error);
-      }
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          try {
+            await fetch(`/api/user/${id}`, { method: 'DELETE' });
+            onDelete(id);
+          } catch (error) {
+            console.error('Error deleting user:', error);
+          }
+
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your file has been deleted.",
+            icon: "success"
+          });
+        }
+      });
     };
   
     return (
@@ -22,12 +44,12 @@ const UserList = ({ users, onDelete }) => {
         <tbody>
           {users.map((user) => (
             <tr key={user._id} className="text-white border-b border-white">
-              <td className="py-2 px-4">{user.name}</td>
+              <td className="py-2 px-4 capitalize">{user.name}</td>
               <td className="py-2 px-4">{user.email}</td>
               <td className="py-2 px-4">{user.shift}</td>
               <td className="py-2 px-4">{user.type}</td>
               <td className="py-2 px-4">
-                <button className="bg-red-500 hover:bg-red-600 text-white py-1 px-2 rounded" onClick={() => handleDelete(user._id)}>Delete</button>
+                <Button handleOnClick={() => handleDelete(user._id)}>Delete</Button>
               </td>
             </tr>
           ))}
